requirements.txt
absl-py==2.3.1
accelerate==1.12.0
action-msgs==2.0.3
action-tutorials-interfaces==0.33.7
action-tutorials-py==0.33.7
actionlib-msgs==5.3.6
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
ament-cmake-test==2.5.4
ament-copyright==0.17.3
ament-cppcheck==0.17.3
ament-cpplint==0.17.3
ament-flake8==0.17.3
ament-index-python==1.8.1
ament-lint==0.17.3
ament-lint-cmake==0.17.3
ament-package==0.16.4
ament-pep257==0.17.3
ament-uncrustify==0.17.3
ament-xmllint==0.17.3
angles==1.16.1
anyio==4.12.1
async-timeout==5.0.1
attrs==25.4.0
av==16.1.0
blinker==1.9.0
builtin-interfaces==2.0.3
certifi==2026.1.4
charset-normalizer==3.4.4
chumpy==0.70
click==8.3.1
cmeel==0.57.3
composition-interfaces==2.0.3
contourpy==1.3.2
cv-bridge==4.1.0
cycler==0.12.1
daqp==0.7.2
datasets==3.6.0
demo-nodes-py==0.33.7
diagnostic-msgs==5.3.6
diagnostic-updater==4.2.6
dill==0.3.8
domain-coordinator==0.12.0
etils==1.13.0
evdev==1.9.2
example-interfaces==0.12.0
examples-rclpy-executors==0.19.7
examples-rclpy-minimal-action-client==0.19.7
examples-rclpy-minimal-action-server==0.19.7
examples-rclpy-minimal-client==0.19.7
examples-rclpy-minimal-publisher==0.19.7
examples-rclpy-minimal-service==0.19.7
examples-rclpy-minimal-subscriber==0.19.7
exceptiongroup==1.3.1
filelock==3.20.1
Flask==3.1.2
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2024.6.1
-e git+https://github.com/YanjieZe/GMR.git@151441704dec25d3b729d67f5e7f3618dcb420b4#egg=general_motion_retargeting
geometry-msgs==5.3.6
glfw==2.10.0
h11==0.16.0
h5py==3.15.1
hf-xet==1.2.0
hiredis==3.3.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.3.4
idna==3.11
image-geometry==4.1.0
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
importlib_resources==6.5.2
interactive-markers==2.5.5
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
jsonlines==4.0.0
keyboard==0.13.5
kiwisolver==1.4.9
laser-geometry==2.7.2
launch==3.4.9
launch-ros==0.26.10
launch-testing==3.4.9
launch-testing-ros==0.26.10
launch-xml==3.4.9
launch-yaml==3.4.9
lerobot==0.3.2
lifecycle-msgs==2.0.3
logging-demo==0.33.7
loop-rate-limiters==1.2.0
map-msgs==2.4.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.3.4
mdurl==0.1.2
message-filters==4.11.9
mink==1.0.0
mpmath==1.3.0
msgpack==1.1.2
msgpack-python==0.5.6
mujoco==3.4.0
multidict==6.7.1
multiprocess==0.70.16
natsort==8.4.0
nav-msgs==5.3.6
networkx==3.4.2
numpy==1.22.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
opencv-python==4.12.0.88
osrf-pycommon==2.1.7
packaging==25.0
pandas==2.3.3
pcl-msgs==1.0.0
pendulum-msgs==0.33.7
pillow==12.0.0
propcache==0.4.1
protobuf==6.33.2
proxsuite==0.7.2
psutil==7.2.1
pyarrow==23.0.0
Pygments==2.19.2
pynput==1.8.1
PyOpenGL==3.1.10
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-qt-binding==2.2.2
python-xlib==0.33
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
qpsolvers==4.8.2
qt-dotgraph==2.7.5
qt-gui==2.7.5
qt-gui-cpp==2.7.5
qt-gui-py-common==2.7.5
quality-of-service-demo-py==0.33.7
rcl-interfaces==2.0.3
rclpy==7.1.6
rcutils==6.7.4
redis==7.1.0
regex==2026.1.15
requests==2.32.5
resource-retriever==3.4.4
rich==14.2.0
rmw-dds-common==3.1.0
ros2action==0.32.6
ros2bag==0.26.9
ros2bag-mcap-cli==0.26.9
ros2bag-sqlite3-cli==0.26.9
ros2cli==0.32.6
ros2component==0.32.6
ros2doctor==0.32.6
ros2interface==0.32.6
ros2launch==0.26.10
ros2lifecycle==0.32.6
ros2multicast==0.32.6
ros2node==0.32.6
ros2param==0.32.6
ros2pkg==0.32.6
ros2run==0.32.6
ros2service==0.32.6
ros2topic==0.32.6
rosbag2-interfaces==0.26.9
rosbag2-py==0.26.9
rosgraph-msgs==2.0.3
rosidl-adapter==4.6.6
rosidl-cli==4.6.6
rosidl-cmake==4.6.6
rosidl-generator-c==4.6.6
rosidl-generator-cpp==4.6.6
rosidl-generator-py==0.22.2
rosidl-generator-type-description==4.6.6
rosidl-parser==4.6.6
rosidl-pycommon==4.6.6
rosidl-runtime-py==0.13.1
rosidl-typesupport-c==3.2.2
rosidl-typesupport-cpp==3.2.2
rosidl-typesupport-fastrtps-c==3.6.2
rosidl-typesupport-fastrtps-cpp==3.6.2
rosidl-typesupport-introspection-c==4.6.6
rosidl-typesupport-introspection-cpp==4.6.6
rpyutils==0.4.2
rqt-action==2.2.0
rqt-bag==1.5.5
rqt-bag-plugins==1.5.5
rqt-console==2.2.1
rqt-graph==1.5.5
rqt-gui==1.6.2
rqt-gui-py==1.6.2
rqt-msg==1.5.1
rqt-plot==1.4.4
rqt-publisher==1.7.2
rqt-py-common==1.6.2
rqt-py-console==1.2.3
rqt-reconfigure==1.6.2
rqt-service-caller==1.2.1
rqt-shell==1.2.2
rqt-srv==1.2.2
rqt-topic==1.7.4
safetensors==0.7.0
scipy==1.15.3
sensor-msgs==5.3.6
sensor-msgs-py==5.3.6
service-msgs==2.0.3
shape-msgs==5.3.6
shellingham==1.5.4
six==1.17.0
smplx @ file:///home/wsj/Desktop/code/github/GMR/dependments/smplx-main
sros2==0.13.4
statistics-msgs==2.0.3
std-msgs==5.3.6
std-srvs==5.3.6
stereo-msgs==5.3.6
sympy==1.14.0
teleop-twist-keyboard==2.4.1
tf2-geometry-msgs==0.36.16
tf2-kdl==0.36.16
tf2-msgs==0.36.16
tf2-py==0.36.16
tf2-ros-py==0.36.16
tf2-sensor-msgs==0.36.16
tf2-tools==0.36.16
tokenizers==0.22.2
tomli==2.3.0
topic-monitor==0.33.7
torch==2.8.0+cu128
torchaudio==2.8.0+cu128
torchcodec==0.10.0
torchvision==0.23.0+cu128
tqdm==4.67.1
trajectory-msgs==5.3.6
transformers==5.0.0
triton==3.4.0
trl==0.27.2
turtlesim==1.8.3
type-description-interfaces==2.0.3
typer-slim==0.21.1
typing_extensions==4.15.0
tzdata==2025.3
unique-identifier-msgs==2.5.0
urllib3==2.6.3
visualization-msgs==5.3.6
Werkzeug==3.1.5
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zmq==0.0.0
